flet_app/requirements.txt
flet
PyYAML
pytest>=7.0
pytest-mock>=3.10
pytest-timeout>=2.1
pytest-cov>=4.0